+++ app/src/main/java/ch/beerpro/presentation/profile/myfridge/OnFridgeInteractionListener.java
@@ -1,0 +1,9 @@
+package ch.beerpro.presentation.profile.myfridge;
+
+import android.widget.ImageView;
+
+import ch.beerpro.domain.models.Beer;
+import ch.beerpro.presentation.utils.OnDefaultBeerInteractionListener;
+
+interface OnFridgeInteractionListener extends OnDefaultBeerInteractionListener {
+}
