+++ app/src/main/java/ch/beerpro/presentation/profile/mywishlist/OnWishlistItemInteractionListener.java
@@ -2,10 +2,9 @@ package ch.beerpro.presentation.profile.mywishlist;
 
 import android.widget.ImageView;
 import ch.beerpro.domain.models.Beer;
+import ch.beerpro.presentation.utils.OnDefaultBeerInteractionListener;
 
-public interface OnWishlistItemInteractionListener {
-
-    void onMoreClickedListener(ImageView photo, Beer beer);
+public interface OnWishlistItemInteractionListener extends OnDefaultBeerInteractionListener {
 
     void onWishClickedListener(Beer beer);
 }
