+++ app/src/main/java/ch/beerpro/presentation/profile/mywishlist/WishlistRecyclerViewAdapter.java
@@ -8,6 +8,7 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.RatingBar;
 import android.widget.TextView;
+
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.ListAdapter;
@@ -16,9 +17,12 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import ch.beerpro.GlideApp;
 import ch.beerpro.R;
+import ch.beerpro.presentation.utils.DefaultBeerViewHolder;
 import ch.beerpro.presentation.utils.EntityPairDiffItemCallback;
 import ch.beerpro.domain.models.Beer;
 import ch.beerpro.domain.models.Wish;
+import ch.beerpro.presentation.utils.OnDefaultBeerInteractionListener;
+
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
 
@@ -52,26 +56,7 @@ public class WishlistRecyclerViewAdapter extends ListAdapter<Pair<Wish, Beer>, W
         holder.bind(item.first, item.second, listener);
     }
 
-    class ViewHolder extends RecyclerView.ViewHolder {
-
-        @BindView(R.id.name)
-        TextView name;
-
-        @BindView(R.id.manufacturer)
-        TextView manufacturer;
-
-        @BindView(R.id.category)
-        TextView category;
-
-        @BindView(R.id.photo)
-        ImageView photo;
-
-        @BindView(R.id.ratingBar)
-        RatingBar ratingBar;
-
-        @BindView(R.id.numRatings)
-        TextView numRatings;
-
+    class ViewHolder extends DefaultBeerViewHolder {
         @BindView(R.id.addedAt)
         TextView addedAt;
 
@@ -84,17 +69,7 @@ public class WishlistRecyclerViewAdapter extends ListAdapter<Pair<Wish, Beer>, W
         }
 
         void bind(Wish wish, Beer item, OnWishlistItemInteractionListener listener) {
-            name.setText(item.getName());
-            manufacturer.setText(item.getManufacturer());
-            category.setText(item.getCategory());
-            name.setText(item.getName());
-            GlideApp.with(itemView).load(item.getPhoto()).apply(new RequestOptions().override(240, 240).centerInside())
-                    .into(photo);
-            ratingBar.setNumStars(5);
-            ratingBar.setRating(item.getAvgRating());
-            numRatings.setText(itemView.getResources().getString(R.string.fmt_num_ratings, item.getNumRatings()));
-            itemView.setOnClickListener(v -> listener.onMoreClickedListener(photo, item));
-
+            super.bind(item, listener);
             String formattedDate =
                     DateFormat.getDateTimeInstance(DateFormat.FULL, DateFormat.SHORT).format(wish.getAddedAt());
             addedAt.setText(formattedDate);
